refactor botorch utils into separate module
